export bulk export relationships logic
so that functionality built on top of the
Datastore API can also perform BulkExport
